refactor: split the message sending logic from content script
